:bug: bug: fix movie trailer load multi-times in BannerItem
