+++ app/src/components/media/MediaList.tsx
@@ -2,12 +2,10 @@ import * as React from 'react';
 import { Text, Button, Row } from '@nextui-org/react';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Swiper as SwiperClass } from 'swiper/types';
 
 import { IMedia } from '~/services/tmdb/tmdb.types';
 import ChevronRightIcon from '~/src/assets/icons/ChevronRightIcon.js';
 import ChevronLeftIcon from '~/src/assets/icons/ChevronLeftIcon.js';
-import { Trailer } from '~/src/components/elements/modal/WatchTrailerModal';
 import { MediaListTable, MediaListCard, MediaListBanner, MediaListGrid } from './list';
 import Filter from '../elements/filter/Filter';
 
@@ -31,12 +29,6 @@ interface IMediaListProps {
   onClickViewMore?: () => void;
   cardType?: 'media' | 'similar-movie' | 'similar-tv';
   navigationButtons?: boolean;
-  handleSlideChangeTransitionEnd?: (swiper: SwiperClass) => void;
-  handleSlideChangeTransitionStart?: (swiper: SwiperClass) => void;
-  handleTouchMove?: (swiper: SwiperClass, e: MouseEvent | TouchEvent | PointerEvent) => void;
-  setShowTrailer?: React.Dispatch<React.SetStateAction<boolean>>;
-  showTrailer?: boolean;
-  trailer?: Trailer;
 }
 
 const MediaList = (props: IMediaListProps) => {
@@ -52,12 +44,6 @@ const MediaList = (props: IMediaListProps) => {
     onClickViewMore,
     cardType,
     navigationButtons,
-    handleSlideChangeTransitionEnd,
-    handleSlideChangeTransitionStart,
-    handleTouchMove,
-    setShowTrailer,
-    showTrailer,
-    trailer,
   } = props;
   let { listType } = props;
 
@@ -89,19 +75,7 @@ const MediaList = (props: IMediaListProps) => {
       list = <MediaListTable items={items} />;
       break;
     case 'slider-banner':
-      list = (
-        <MediaListBanner
-          items={items}
-          handleSlideChangeTransitionEnd={handleSlideChangeTransitionEnd}
-          handleSlideChangeTransitionStart={handleSlideChangeTransitionStart}
-          handleTouchMove={handleTouchMove}
-          genresMovie={genresMovie}
-          genresTv={genresTv}
-          setShowTrailer={setShowTrailer}
-          showTrailer={showTrailer}
-          trailer={trailer}
-        />
-      );
+      list = <MediaListBanner items={items} genresMovie={genresMovie} genresTv={genresTv} />;
       break;
     case 'slider-card':
       list = (
